Adding initial structure for main page This takes care of adding a new path and the initial structure for the homepage which will provide a view of all petitions.
