--- src/main/java/com/example/demo/DemoApplication.java
@@ -1,18 +1,0 @@
-package com.example.demo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import java.util.*;
-
-@SpringBootApplication
-public class DemoApplication {
-
-	public static void main(String[] args) {
-		SpringApplication app = new SpringApplication(DemoApplication.class);
-
-		app.setDefaultProperties(Collections
-				.singletonMap("server.port", "9090"));
-		app.run(args);
-	}
-
-}
