+++ src/main/resources/templates/petitions.html
@@ -1,0 +1,9 @@
+<!doctype html>
+<html>
+    <head>
+        <title> Petition Platform </title>
+    </head>
+    <body>
+        <p> This is a petition website. </p>
+    </body>
+</html>
