+++ src/test/java/com/example/demo/PetitionApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class DemoApplicationTests {
+class PetitionApplicationTests {
 
 	@Test
 	void contextLoads() {
